Update all extractors to return int timestamps, remove unnecessary type unions and casts

+++ custom_components/haeo/data/loader/extractors/aemo_nem.py
@@ -63,10 +63,11 @@ class Parser:
         )
 
     @staticmethod
-    def extract(state: AemoNemState) -> tuple[Sequence[tuple[float, float]], str, SensorDeviceClass]:
+    def extract(state: AemoNemState) -> tuple[Sequence[tuple[int, float]], str, SensorDeviceClass]:
         """Extract forecast data from AEMO energy market format."""
-        parsed: list[tuple[float, float]] = [
-            (parse_datetime_to_timestamp(item["start_time"]), item["price"]) for item in state.attributes["forecast"]
+        parsed: list[tuple[int, float]] = [
+            (int(parse_datetime_to_timestamp(item["start_time"])), item["price"])
+            for item in state.attributes["forecast"]
         ]
         parsed.sort(key=lambda x: x[0])
         return parsed, Parser.UNIT, Parser.DEVICE_CLASS
