+++ custom_components/haeo/data/loader/extractors/amber2mqtt.py
@@ -76,10 +76,10 @@ class Parser:
         )
 
     @staticmethod
-    def _round_to_minute(timestamp: str | datetime) -> float:
+    def _round_to_minute(timestamp: str | datetime) -> int:
         """Round timestamp to nearest minute (Amber provides times 1 second into each period)."""
         raw = float(parse_datetime_to_timestamp(timestamp))
-        return round(raw / 60.0) * 60.0
+        return int(round(raw / 60.0) * 60.0)
 
     @staticmethod
     def extract(state: Amber2MqttState) -> tuple[Sequence[tuple[int, float]], str, SensorDeviceClass]:
@@ -100,8 +100,8 @@ class Parser:
         is_feedin = state.attributes.get("channel_type") == "feedin"
 
         for item in forecasts:
-            start = int(Parser._round_to_minute(item["start_time"]))
-            end = int(Parser._round_to_minute(item["end_time"]))
+            start = Parser._round_to_minute(item["start_time"])
+            end = Parser._round_to_minute(item["end_time"])
             price = -item["per_kwh"] if is_feedin else item["per_kwh"]
 
             # Emit start of window and end of window with same price
